Csv user import with ely numbers

+++ digiroad2-geo/src/main/scala/fi/liikennevirasto/digiroad2/asset/asset.scala
@@ -6,19 +6,19 @@ case class AssetType(id: Long, assetTypeName: String, geometryType: String)
 case class Asset(id: Long, assetTypeId: Long, lon: Double, lat: Double, roadLinkId: Long,
                  imageIds: Seq[String] = List(), bearing: Option[Int] = None, validityDirection: Option[Int] = None,
                  status: Option[String] = None, readOnly: Boolean = true,
-                 municipalityNumber: Option[Long] = None, validityPeriod: Option[String] = None)
+                 municipalityNumber: Option[Int] = None, validityPeriod: Option[String] = None)
 
 case class AssetWithProperties(id: Long, assetTypeId: Long, lon: Double, lat: Double, roadLinkId: Long,
                  imageIds: Seq[String] = List(), bearing: Option[Int] = None, validityDirection: Option[Int] = None,
                  status: Option[String] = None, readOnly: Boolean = true,
-                 municipalityNumber: Option[Long] = None,
+                 municipalityNumber: Option[Int] = None,
                  propertyData: Seq[Property] = List(), validityPeriod: Option[String] = None)
 
 case class SimpleProperty(id: String, values: Seq[PropertyValue])
 case class Property(propertyId: String, propertyName: String, propertyType: String, required: Boolean = false, values: Seq[PropertyValue])
 case class PropertyValue(propertyValue: Long, propertyDisplayValue: String, imageId: String = null)
 case class EnumeratedPropertyValue(propertyId: String, propertyName: String, propertyType: String, required: Boolean = false, values: Seq[PropertyValue])
-case class RoadLink(id: Long, lonLat: Seq[(Double, Double)], endDate: Option[LocalDate] = None, municipalityNumber: Long)
+case class RoadLink(id: Long, lonLat: Seq[(Double, Double)], endDate: Option[LocalDate] = None, municipalityNumber: Int)
 
 object PropertyTypes {
   val SingleChoice = "single_choice"
